perubahan alur request activity

+++ database/migrations/2023_03_13_132046_create_cargo_items_table.php
@@ -20,8 +20,8 @@ class CreateCargoItemsTable extends Migration
          $table->string('desc');
          $table->smallInteger('qty');
          $table->string('unit')->nullable();
-         $table->smallInteger('size')->nullable();
-         $table->smallInteger('weight')->nullable();
+         $table->decimal('size', 6, 2)->nullable();
+         $table->decimal('weight', 6, 2)->nullable();
          $table->timestamps();
       });
    }
